+++ packages/app-store/office365_video/lib/VideoApiAdapter.ts
