+++ packages/app-store/office365_video/lib/index.ts
