+++ packages/app-store/office365_video/lib/locationOption.ts
