+++ packages/app-store/office365_video/package.json
