+++ packages/app-store/office365_video/utils.ts
