+++ packages/lib/location.ts
@@ -8,5 +8,5 @@ export enum LocationType {
   Jitsi = "integrations:jitsi",
   Huddle01 = "integrations:huddle01",
   Tandem = "integrations:tandem",
-  Teams = "integrations:msteams",
+  Teams = "integrations:office365_calendar",
 }
